refactor: 디렉토리구조 변경으로 인한 파일명 통합
Admin~~ -> A~~
