+++ server/src/main/java/com/genesisairport/reservation/controller/admin/AReservationController.java
@@ -1,7 +1,9 @@
 package com.genesisairport.reservation.controller.admin;
 
 import com.genesisairport.reservation.common.enums.ResponseCode;
+import com.genesisairport.reservation.common.exception.GeneralException;
 import com.genesisairport.reservation.common.model.ResponseDto;
+import com.genesisairport.reservation.request.AdminRequest;
 import com.genesisairport.reservation.service.S3Service;
 import com.genesisairport.reservation.service.admin.AReservationService;
 import lombok.RequiredArgsConstructor;
@@ -27,6 +29,44 @@ public class AReservationController {
                                                    @RequestPart("image") MultipartFile image) throws IOException {
         String imageUrl = s3Service.saveFile(image);
         aReservationService.addMaintenanceImage(id, status, imageUrl);
+        return new ResponseEntity<>(ResponseDto.of(true, ResponseCode.OK), HttpStatus.OK);
+    }
+
+    @PostMapping("/progress")
+    public ResponseEntity registerStage(
+            @RequestBody AdminRequest.StageInfo requestBody) {
+
+        if (requestBody.getProgress() == null) {
+            throw new GeneralException(ResponseCode.INTERNAL_ERROR, "유효하지 않은 진행단계입니다.");
+        }
+
+        aReservationService.saveStage(requestBody);
+        return new ResponseEntity<>(ResponseDto.of(true, ResponseCode.OK), HttpStatus.OK);
+    }
+
+    @DeleteMapping("/progress")
+    public ResponseEntity deleteStage(
+            @RequestBody AdminRequest.StageInfo requestBody) {
+
+        if (requestBody.getProgress() == null) {
+            throw new GeneralException(ResponseCode.INTERNAL_ERROR, "유효하지 않은 진행단계입니다.");
+        }
+
+        aReservationService.deleteStage(requestBody);
+        return new ResponseEntity<>(ResponseDto.of(true, ResponseCode.OK), HttpStatus.OK);
+    }
+
+    @PutMapping("/comment")
+    public ResponseEntity updateComment(@RequestBody AdminRequest.CommentInfo requestBody) {
+
+        if (requestBody.getReservationId() == null)
+            throw new GeneralException(ResponseCode.INTERNAL_ERROR, "예약 Id를 받아오지 못했습니다.");
+
+        if (requestBody.getComment() == null)
+            throw new GeneralException(ResponseCode.INTERNAL_ERROR, "코멘트를 받아오지 못했습니다.");
+
+        aReservationService.updateComment(requestBody);
+
         return new ResponseEntity<>(ResponseDto.of(true, ResponseCode.OK), HttpStatus.OK);
     }
 }
